Move all presentation to Document container
